+++ resolwe/flow/tests/test_api.py
@@ -252,6 +252,7 @@ class TestDuplicate(TransactionTestCase):
         force_authenticate(request, self.contributor)
         response = self.entity_duplicate_viewset(request)
         task = BackgroundTask.objects.get(pk=response.data["id"])
+        task.result()
         duplicate = Entity.objects.get(id__in=task.result())
 
         self.assertTrue(task.has_permission(Permission.VIEW, self.contributor))
@@ -278,7 +279,10 @@ class TestDuplicate(TransactionTestCase):
             type="STRING",
         )
         AnnotationValue.objects.create(
-            field=annotation_field1, entity=entity, value="test"
+            field=annotation_field1,
+            entity=entity,
+            value="test",
+            contributor=self.contributor,
         )
         entity.collection.annotation_fields.add(annotation_field1)
         request = factory.post(
@@ -289,7 +293,7 @@ class TestDuplicate(TransactionTestCase):
         force_authenticate(request, self.contributor)
         response = self.entity_duplicate_viewset(request)
         task = BackgroundTask.objects.get(pk=response.data["id"])
-        result = task.result(final_statuses=["OK"])
+        result = task.result()
         duplicate = Entity.objects.get(id__in=result)
         self.assertTrue(task.has_permission(Permission.VIEW, self.contributor))
         self.assertEqual(len(result), 1)
@@ -877,8 +881,10 @@ class TestDataViewSetCase(TestDataViewSetCaseMixin, TestCase):
         self.assertEqual(data.tags, collection.tags)
 
     def test_process_is_active(self):
-        # Do not allow creating data of inactive processes
-        Process.objects.filter(slug="test-process").update(is_active=False)
+        # Do not allow creating data of inactive processes.
+        # The all_objects is important, since window function (used by default manager)
+        # is not compatible with the update statement.
+        Process.all_objects.filter(slug="test-process").update(is_active=False)
         data = {"process": {"slug": "test-process"}}
         request = factory.post("/", data, format="json")
         force_authenticate(request, self.contributor)