+++ resolwe/process/tests/test_python_process.py
@@ -163,7 +163,9 @@ class PythonProcessTest(ProcessTestCase):
         # Now try reading and updating existing annotations.
         entity = Entity.objects.create(name="Entity", contributor=self.contributor)
         entity.set_permission(Permission.EDIT, self.contributor)
-        AnnotationValue.objects.create(entity=entity, field=age_field, value=42)
+        AnnotationValue.objects.create(
+            entity=entity, field=age_field, value=42, contributor=self.contributor
+        )
         process = Process.objects.get(
             slug="test-python-process-update-entity-annotations-v2"
         )